django/analytics/requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.4.0
Django==2.2.6
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Markdown==3.1.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
